style(open-api-page): gather types for Typescript conversion
